object-backend/pyproject.toml
[project]
name = "object-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.3",
    "djangorestframework>=3.15.2",
    "pillow>=11.0.0",
    "ruff>=0.7.3",
]
